# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AtomQuest Goal Setting & Tracking Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List goals (filtered by role context)
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: approvalStatus
          in: query
          schema:
            type: string
        - name: quarter
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
        - name: thrustArea
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a new goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Goal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    get:
      operationId: getGoal
      tags: [goals]
      summary: Get a goal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Goal found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "404":
          description: Not found
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a draft goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /goals/{id}/submit:
    post:
      operationId: submitGoal
      tags: [goals]
      summary: Submit a draft goal for approval
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}/approve:
    post:
      operationId: approveGoal
      tags: [goals]
      summary: Approve a goal (manager)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalReviewInput"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}/reject:
    post:
      operationId: rejectGoal
      tags: [goals]
      summary: Reject a goal (manager)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalReviewInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}/return:
    post:
      operationId: returnGoal
      tags: [goals]
      summary: Return goal for rework (manager)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalReviewInput"
      responses:
        "200":
          description: Returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}/unlock:
    post:
      operationId: unlockGoal
      tags: [goals]
      summary: Unlock a locked goal (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unlocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}/quarterly-update:
    post:
      operationId: addQuarterlyUpdate
      tags: [goals]
      summary: Add or update quarterly achievement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuarterlyUpdateInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}/comments:
    post:
      operationId: addGoalComment
      tags: [goals]
      summary: Add a comment to a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "200":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"

  # ─── SHARED GOALS ───────────────────────────────────────
...
components:
  schemas:
...
    Goal:
      type: object
      required: [id, employeeId, thrustArea, goalTitle, goalDescription, unitOfMeasurement, target, weightage, approvalStatus, lockStatus, createdAt]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        thrustArea:
          type: string
        goalTitle:
          type: string
        goalDescription:
          type: string
        unitOfMeasurement:
          type: string
          enum: [numeric, percentage, timeline, zero_based]
        scoreType:
          type: string
          enum: [higher_is_better, lower_is_better]
        target:
          type: number
        deadline:
          type: ["string", "null"]
          format: date
        weightage:
          type: number
        achievement:
          type: ["number", "null"]
        progressScore:
          type: ["number", "null"]
        approvalStatus:
          type: string
          enum: [draft, pending, approved, rejected, returned]
        lockStatus:
          type: boolean
        quarter:
          type: ["string", "null"]
        quarterlyUpdates:
          type: array
          items:
            $ref: "#/components/schemas/QuarterlyUpdate"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        isShared:
          type: boolean
        sharedGoalId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    GoalInput:
      type: object
      required: [thrustArea, goalTitle, goalDescription, unitOfMeasurement, target, weightage]
      properties:
        thrustArea:
          type: string
        goalTitle:
          type: string
        goalDescription:
          type: string
        unitOfMeasurement:
          type: string
          enum: [numeric, percentage, timeline, zero_based]
        scoreType:
          type: string
          enum: [higher_is_better, lower_is_better]
        target:
          type: number
        deadline:
          type: string
          format: date
        weightage:
          type: number
        quarter:
          type: string
    GoalUpdate:
      type: object
      properties:
        thrustArea:
          type: string
        goalTitle:
          type: string
        goalDescription:
          type: string
        unitOfMeasurement:
          type: string
          enum: [numeric, percentage, timeline, zero_based]
        scoreType:
          type: string
          enum: [higher_is_better, lower_is_better]
        target:
          type: number
        deadline:
          type: string
          format: date
        weightage:
          type: number
        achievement:
          type: number
    GoalReviewInput:
      type: object
      required: [comment]
      properties:
        comment:
          type: string
        targetOverride:
          type: ["number", "null"]
        weightageOverride:
          type: ["number", "null"]
    QuarterlyUpdate:
      type: object
      required: [quarter, achievement, progressStatus, updatedAt]
      properties:
        quarter:
          type: string
        achievement:
          type: number
        progressStatus:
          type: string
          enum: [not_started, on_track, completed, behind]
        progressScore:
          type: ["number", "null"]
        managerComment:
          type: ["string", "null"]
        updatedAt:
          type: string
          format: date-time
    QuarterlyUpdateInput:
      type: object
      required: [quarter, achievement, progressStatus]
      properties:
        quarter:
          type: string
        achievement:
          type: number
        progressStatus:
          type: string
          enum: [not_started, on_track, completed, behind]
    Comment:
      type: object
      required: [id, authorId, authorName, role, text, createdAt]
      properties:
        id:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        role:
          type: string
        text:
          type: string
        createdAt:
          type: string
          format: date-time
    CommentInput:
      type: object
      required: [text]
      properties:
        text:
          type: string

    # ── Shared Goal ─────────────────────────────────────
...